Improved IP ban checks
