+++ routes/accounts.js
@@ -3,7 +3,7 @@ const Account = require('../models/account');
 const Profile = require('../models/profile/index');
 const BannedIP = require('../models/bannedIP');
 const EightEightCounter = require('../models/eightEightCounter');
-const { ipbansNotEnforced, accountCreationDisabled, verifyBypass, consumeBypass } = require('./socket/models');
+const { ipbansNotEnforced, accountCreationDisabled, verifyBypass, consumeBypass, testIP } = require('./socket/models');
 const { verifyRoutes, setVerify } = require('./verification');
 const blacklistedWords = require('../iso/blacklistwords');
 const bannedEmails = require('../utils/disposibleEmails');
@@ -20,51 +20,6 @@ const ensureAuthenticated = (req, res, next) => {
 		return next();
 	}
 	res.redirect('/');
-};
-
-let banCache = null;
-setInterval(() => {
-	// Fetches the list of banned IPs every 5 seconds, to prevent hammering the DB on restarts as people log in.
-	BannedIP.find({}, (err, ips) => {
-		if (err) console.log(err);
-		else banCache = ips;
-	});
-}, 5000);
-// There's a mountain of "new" type bans.
-const unbanTime = new Date() - 64800000;
-BannedIP.deleteMany({ type: 'new', bannedDate: { $lte: unbanTime } }, (err, r) => {
-	if (err) throw err;
-	BannedIP.find({}, (err, ips) => {
-		if (err) throw err;
-		banCache = ips;
-	});
-});
-const banLength = {
-	small: 18 * 60 * 60 * 1000, // 18 hours
-	new: 18 * 60 * 60 * 1000, // 18 hours
-	tiny: 1 * 60 * 60 * 1000, // 1 hour
-	big: 7 * 24 * 60 * 60 * 1000 // 7 days
-};
-const testIP = (IP, callback) => {
-	if (!IP) callback('Bad IP!');
-	else if (!banCache || !banCache.filter) callback('nocache');
-	else {
-		const ips = banCache.filter(i => i.ip == IP);
-		let date;
-		let unbannedTime;
-		const ip = ips[ips.length - 1];
-
-		if (ip) {
-			date = new Date().getTime();
-			unbannedTime = ip.bannedDate.getTime() + (banLength[ip.type] || banLength.big);
-		}
-
-		if (ip && unbannedTime > date && !ipbansNotEnforced.status && process.env.NODE_ENV === 'production') {
-			callback(ip.type);
-		} else {
-			callback(null);
-		}
-	}
 };
 
 const emailRegex = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/;
@@ -138,8 +93,9 @@ module.exports = () => {
 				hasBypass = true;
 			}
 		}
-		const signupIP =
-			req.headers['x-real-ip'] || req.headers['X-Real-IP'] || req.headers['X-Forwarded-For'] || req.headers['x-forwarded-for'] || req.connection.remoteAddress;
+		const signupIP = expandAndSimplify(
+			req.headers['x-real-ip'] || req.headers['X-Real-IP'] || req.headers['X-Forwarded-For'] || req.headers['x-forwarded-for'] || req.connection.remoteAddress
+		);
 		const save = {
 			username,
 			gameSettings: {
@@ -284,11 +240,13 @@ module.exports = () => {
 		'/account/signin',
 		(req, res, next) => {
 			testIP(
-				req.headers['x-real-ip'] ||
-					req.headers['X-Real-IP'] ||
-					req.headers['X-Forwarded-For'] ||
-					req.headers['x-forwarded-for'] ||
-					req.connection.remoteAddress,
+				expandAndSimplify(
+					req.headers['x-real-ip'] ||
+						req.headers['X-Real-IP'] ||
+						req.headers['X-Forwarded-For'] ||
+						req.headers['x-forwarded-for'] ||
+						req.connection.remoteAddress
+				),
 				banType => {
 					if (banType && banType != 'new') {
 						if (banType == 'nocache') res.status(403).json({ message: 'The server is still getting its bearings, try again in a few moments.' });
